+++ src/gui/console.d
@@ -70,11 +70,11 @@ class ConsoleEditLine : EditLine {
 
 //Replace the text between start and end by text
 //operation: newline = line[0..start] ~ text ~ line[end..$];
-alias void delegate(int start, int end, cstring text) EditDelegate;
+alias void delegate(size_t start, size_t end, cstring text) EditDelegate;
 
 //line is the current line, cursor1+cursor2 are the selection/cursor-
 //  position, and edit can be used to change the text
-alias void delegate(cstring line, int cursor1, int cursor2, EditDelegate edit)
+alias void delegate(cstring line, size_t cursor1, size_t cursor2, EditDelegate edit)
     TabCompleteDelegate;
 
 //a generic console (basically just a LogWindow over an EditLine)
